[MOVE] 로직 순서 및 파일 이동
