+++ app/src/main/java/com/example/allclear/timetable/maketimetable/generalelective/SelectGeneralElectiveActivity.java
@@ -21,8 +21,8 @@ import com.example.allclear.data.response.TimeTableResponseDto;
 import com.example.allclear.databinding.ActivitySelectGeneralElectiveBinding;
 import com.example.allclear.databinding.SpinnerCustomBinding;
 import com.example.allclear.schedule.AdapterSpinner;
-import com.example.allclear.timetable.maketimetable.EssentialSubjectActivity;
 import com.example.allclear.timetable.maketimetable.MakeTimeTableAdapter;
+import com.example.allclear.timetable.maketimetable.selfadd.SelfAddOneActivity;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -53,7 +53,7 @@ public class SelectGeneralElectiveActivity extends AppCompatActivity {
         initNextClickListener();
         initBackClickListener();
         getGeneralElectiveList();
-        setGeneralSpinner();
+     //   setGeneralSpinner();
 
     }
 
@@ -80,7 +80,7 @@ public class SelectGeneralElectiveActivity extends AppCompatActivity {
                     @Override
                     public void onResponse(@NonNull Call<TimeTableResponseDto> call, @NonNull Response<TimeTableResponseDto> response) {
                         if (response.isSuccessful()) {
-                            Intent intent = new Intent(SelectGeneralElectiveActivity.this, EssentialSubjectActivity.class);
+                            Intent intent = new Intent(SelectGeneralElectiveActivity.this, SelfAddOneActivity.class);
                             startActivity(intent);
                         }
                     }
