+++ app/src/main/res/layout/activity_select_major.xml
@@ -66,7 +66,7 @@
                     android:id="@+id/rv_select_major"
                     android:layout_width="0dp"
                     android:layout_height="0dp"
-                    android:layout_marginTop="10dp"
+                    android:layout_marginTop="20dp"
                     android:scrollbarFadeDuration="0"
                     android:scrollbarSize="5dp"
                     android:scrollbarThumbVertical="@android:color/darker_gray"
