+++ app/src/main/res/layout/activity_self_add_two.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".timetable.maketimetable.SelfAddTwoActivity">
+    tools:context=".timetable.maketimetable.selfadd.SelfAddTwoActivity">
 
     <TextView
         android:id="@+id/tv_title"
